# contract/openapi.json
{
  "components" : {
    "schemas" : {
...
      "AiUsageSummaryItem" : {
        "properties" : {
          "estimatedCostUsd" : {
            "type" : "number"
          },
          "feature" : {
            "type" : "string"
          },
          "inputTokens" : {
            "format" : "int64",
            "type" : "integer"
          },
          "model" : {
            "type" : "string"
          },
          "outputTokens" : {
            "format" : "int64",
            "type" : "integer"
          },
          "requests" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "BucketAssignmentRequest" : {
        "properties" : {
          "assignments" : {
            "items" : {
              "$ref" : "#/components/schemas/Item"
            },
            "type" : "array"
          }
        },
        "required" : [ "assignments" ],
        "type" : "object"
      },
      "BucketSummary" : {
        "properties" : {
          "bucket" : {
            "enum" : [ "NEEDS", "WANTS", "SAVINGS" ],
            "type" : "string"
          },
          "percent" : {
            "format" : "int32",
            "type" : "integer"
          },
          "percentUsed" : {
            "format" : "double",
            "type" : "number"
          },
          "remaining" : {
            "type" : "number"
          },
          "spent" : {
            "type" : "number"
          },
          "target" : {
            "type" : "number"
          }
        },
        "type" : "object"
      },
      "BucketSummaryV2" : {
        "properties" : {
          "bucket" : {
            "enum" : [ "NEEDS", "WANTS", "SAVINGS" ],
            "type" : "string"
          },
          "percent" : {
            "format" : "int32",
            "type" : "integer"
          },
          "percentUsed" : {
            "format" : "double",
            "type" : "number"
          },
          "remaining" : {
            "format" : "int64",
            "type" : "integer"
          },
          "spent" : {
            "format" : "int64",
            "type" : "integer"
          },
          "target" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "BudgetRuleEnabledRequest" : {
        "properties" : {
          "enabled" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
      "BudgetRuleRequest" : {
        "properties" : {
          "monthlyIncome" : {
            "minimum" : 0.0,
            "type" : "number"
          },
          "needsPct" : {
            "format" : "int32",
            "type" : "integer"
          },
          "ruleType" : {
            "enum" : [ "FIFTY_THIRTY_TWENTY", "SEVENTY_TWENTY_TEN", "CUSTOM" ],
            "type" : "string"
          },
          "savingsPct" : {
            "format" : "int32",
            "type" : "integer"
          },
          "wantsPct" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "monthlyIncome", "ruleType" ],
        "type" : "object"
      },
      "BudgetRuleRequestV2" : {
        "properties" : {
          "monthlyIncome" : {
            "format" : "int64",
            "minimum" : 0,
            "type" : "integer"
          },
          "needsPct" : {
            "format" : "int32",
            "type" : "integer"
          },
          "ruleType" : {
            "enum" : [ "FIFTY_THIRTY_TWENTY", "SEVENTY_TWENTY_TEN", "CUSTOM" ],
            "type" : "string"
          },
          "savingsPct" : {
            "format" : "int32",
            "type" : "integer"
          },
          "wantsPct" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "monthlyIncome", "ruleType" ],
        "type" : "object"
      },
      "BudgetRuleResponse" : {
        "properties" : {
          "enabled" : {
            "type" : "boolean"
          },
          "monthlyIncome" : {
            "type" : "number"
          },
          "needsPct" : {
            "format" : "int32",
            "type" : "integer"
          },
          "ruleType" : {
            "enum" : [ "FIFTY_THIRTY_TWENTY", "SEVENTY_TWENTY_TEN", "CUSTOM" ],
            "type" : "string"
          },
          "savingsPct" : {
            "format" : "int32",
            "type" : "integer"
          },
          "wantsPct" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "BudgetRuleResponseV2" : {
        "properties" : {
          "enabled" : {
            "type" : "boolean"
          },
          "monthlyIncome" : {
            "format" : "int64",
            "type" : "integer"
          },
          "needsPct" : {
            "format" : "int32",
            "type" : "integer"
          },
          "ruleType" : {
            "enum" : [ "FIFTY_THIRTY_TWENTY", "SEVENTY_TWENTY_TEN", "CUSTOM" ],
            "type" : "string"
          },
          "savingsPct" : {
            "format" : "int32",
            "type" : "integer"
          },
          "wantsPct" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "BudgetRuleSummaryResponse" : {
        "properties" : {
          "buckets" : {
            "items" : {
              "$ref" : "#/components/schemas/BucketSummary"
            },
            "type" : "array"
          },
          "month" : {
            "type" : "string"
          },
          "monthlyIncome" : {
            "type" : "number"
          },
          "ruleType" : {
            "enum" : [ "FIFTY_THIRTY_TWENTY", "SEVENTY_TWENTY_TEN", "CUSTOM" ],
            "type" : "string"
          },
          "unassignedSpent" : {
            "type" : "number"
          }
        },
        "type" : "object"
      },
      "BudgetRuleSummaryResponseV2" : {
        "properties" : {
          "buckets" : {
            "items" : {
              "$ref" : "#/components/schemas/BucketSummaryV2"
            },
            "type" : "array"
          },
          "month" : {
            "type" : "string"
          },
          "monthlyIncome" : {
            "format" : "int64",
            "type" : "integer"
          },
          "ruleType" : {
            "enum" : [ "FIFTY_THIRTY_TWENTY", "SEVENTY_TWENTY_TEN", "CUSTOM" ],
            "type" : "string"
          },
          "unassignedSpent" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "BudgetSummaryItem" : {
        "properties" : {
          "budgeted" : {
            "type" : "number"
          },
          "categoryId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "categoryName" : {
            "type" : "string"
          },
          "percentUsed" : {
            "type" : "number"
          },
          "remaining" : {
            "type" : "number"
          },
          "spent" : {
            "type" : "number"
          },
          "status" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "BudgetSummaryItemV2" : {
        "properties" : {
          "budgeted" : {
            "format" : "int64",
            "type" : "integer"
          },
          "categoryId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "categoryName" : {
            "type" : "string"
          },
          "percentUsed" : {
            "type" : "number"
          },
          "remaining" : {
            "format" : "int64",
            "type" : "integer"
          },
          "spent" : {
            "format" : "int64",
            "type" : "integer"
          },
          "status" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "BudgetSummaryResponse" : {
        "properties" : {
          "dailyAllowance" : {
            "type" : "number"
          },
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/BudgetSummaryItem"
            },
            "type" : "array"
          },
          "month" : {
            "type" : "string"
          },
          "safeToSpend" : {
            "type" : "number"
          },
          "totalBudgeted" : {
            "type" : "number"
          },
          "totalSpent" : {
            "type" : "number"
          }
        },
        "type" : "object"
      },
      "BudgetSummaryResponseV2" : {
        "properties" : {
          "dailyAllowance" : {
            "format" : "int64",
            "type" : "integer"
          },
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/BudgetSummaryItemV2"
            },
            "type" : "array"
          },
          "month" : {
            "type" : "string"
          },
          "safeToSpend" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalBudgeted" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalSpent" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "Item" : {
        "properties" : {
          "bucket" : {
            "enum" : [ "NEEDS", "WANTS", "SAVINGS" ],
            "type" : "string"
          },
          "categoryId" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "categoryId" ],
        "type" : "object"
      },
...
      "MonthSummaryInsightResponse" : {
        "properties" : {
          "month" : {
            "type" : "string"
          },
          "summary" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ObservabilitySummary" : {
        "properties" : {
          "activeUsers24h" : {
            "format" : "int64",
            "type" : "integer"
          },
          "activeUsers7d" : {
            "format" : "int64",
            "type" : "integer"
          },
          "signups30d" : {
            "format" : "int64",
            "type" : "integer"
          },
          "signups7d" : {
            "format" : "int64",
            "type" : "integer"
          },
          "signupsToday" : {
            "format" : "int64",
            "type" : "integer"
          },
          "topUsers30d" : {
            "items" : {
              "$ref" : "#/components/schemas/TopUserItem"
            },
            "type" : "array"
          },
          "totalUsers" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "RecommendationsInsightResponse" : {
        "properties" : {
          "month" : {
            "type" : "string"
          },
          "recommendations" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "TopCategoryInsightResponse" : {
        "properties" : {
          "category" : {
            "type" : "string"
          },
          "month" : {
            "type" : "string"
          },
          "percentOfMonthTotal" : {
            "type" : "number"
          },
          "total" : {
            "type" : "number"
          }
        },
        "type" : "object"
      },
      "TopCategoryInsightResponseV2" : {
        "properties" : {
          "category" : {
            "type" : "string"
          },
          "month" : {
            "type" : "string"
          },
          "percentOfMonthTotal" : {
            "type" : "number"
          },
          "total" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "TopUserItem" : {
        "properties" : {
          "requests" : {
            "format" : "int64",
            "type" : "integer"
          },
          "userId" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "Backend-issued JWT from POST /auth/login, sent as `Authorization: Bearer <token>`.",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.1.0",
  "paths" : {
    "/admin/ai-usage/summary" : {
      "get" : {
        "operationId" : "summary_3",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/AiUsageSummaryItem"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "admin-ai-usage-controller" ]
      }
    },
...
    "/admin/observability/summary" : {
      "get" : {
        "operationId" : "summary_2",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ObservabilitySummary"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "admin-observability-controller" ]
      }
    },
...
    "/ai/insights/month-summary" : {
      "get" : {
        "operationId" : "monthSummary",
        "parameters" : [ {
          "in" : "query",
          "name" : "month",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MonthSummaryInsightResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "ai-insight-controller" ]
      }
    },
    "/ai/insights/recommendations" : {
      "get" : {
        "operationId" : "recommendations",
        "parameters" : [ {
          "in" : "query",
          "name" : "month",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RecommendationsInsightResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "ai-insight-controller" ]
      }
    },
    "/ai/insights/top-category" : {
      "get" : {
        "operationId" : "topCategory",
        "parameters" : [ {
          "in" : "query",
          "name" : "month",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TopCategoryInsightResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "ai-insight-controller" ]
      }
    },
...
    "/api/v2/ai/insights/month-summary" : {
      "get" : {
        "operationId" : "v2MonthSummaryInsight",
        "parameters" : [ {
          "in" : "query",
          "name" : "month",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MonthSummaryInsightResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "ai-insight-v-2-controller" ]
      }
    },
    "/api/v2/ai/insights/recommendations" : {
      "get" : {
        "operationId" : "v2RecommendationsInsight",
        "parameters" : [ {
          "in" : "query",
          "name" : "month",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RecommendationsInsightResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "ai-insight-v-2-controller" ]
      }
    },
    "/api/v2/ai/insights/top-category" : {
      "get" : {
        "operationId" : "v2TopCategoryInsight",
        "parameters" : [ {
          "in" : "query",
          "name" : "month",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TopCategoryInsightResponseV2"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "ai-insight-v-2-controller" ]
      }
    },
...
    "/api/v2/budget-rules" : {
      "get" : {
        "operationId" : "v2GetBudgetRule",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BudgetRuleResponseV2"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "budget-rule-v-2-controller" ]
      },
      "put" : {
        "operationId" : "v2UpsertBudgetRule",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BudgetRuleRequestV2"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BudgetRuleResponseV2"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "budget-rule-v-2-controller" ]
      }
    },
    "/api/v2/budget-rules/buckets" : {
      "put" : {
        "operationId" : "v2AssignBuckets",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BucketAssignmentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "budget-rule-v-2-controller" ]
      }
    },
    "/api/v2/budget-rules/enabled" : {
      "put" : {
        "operationId" : "v2SetBudgetRuleEnabled",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BudgetRuleEnabledRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BudgetRuleResponseV2"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "budget-rule-v-2-controller" ]
      }
    },
    "/api/v2/budget-rules/summary" : {
      "get" : {
        "operationId" : "v2BudgetRuleSummary",
        "parameters" : [ {
          "in" : "query",
          "name" : "month",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BudgetRuleSummaryResponseV2"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "budget-rule-v-2-controller" ]
      }
    },
...
    "/api/v2/budgets/summary" : {
      "get" : {
        "operationId" : "v2BudgetSummary",
        "parameters" : [ {
          "in" : "query",
          "name" : "month",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BudgetSummaryResponseV2"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "budget-v-2-controller" ]
      }
    },
...
    "/budget-rules" : {
      "get" : {
        "operationId" : "get_4",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BudgetRuleResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "budget-rule-controller" ]
      },
      "put" : {
        "operationId" : "upsert",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BudgetRuleRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BudgetRuleResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "budget-rule-controller" ]
      }
    },
    "/budget-rules/buckets" : {
      "put" : {
        "operationId" : "assignBuckets",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BucketAssignmentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "budget-rule-controller" ]
      }
    },
    "/budget-rules/enabled" : {
      "put" : {
        "operationId" : "setEnabled",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BudgetRuleEnabledRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BudgetRuleResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "budget-rule-controller" ]
      }
    },
    "/budget-rules/summary" : {
      "get" : {
        "operationId" : "summary_1",
        "parameters" : [ {
          "in" : "query",
          "name" : "month",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BudgetRuleSummaryResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "budget-rule-controller" ]
      }
    },
...
    "/budgets/summary" : {
      "get" : {
        "operationId" : "summary",
        "parameters" : [ {
          "in" : "query",
          "name" : "month",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BudgetSummaryResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "budget-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ]
}